@@ -27,6 +27,8 @@
 		"@discordjs/builders": "^0.12.0",
 		"@discordjs/collection": "^0.5.0",
 		"@discordjs/rest": "^0.3.0",
+		"@fyko/export-api": "^0.2.0",
+		"@grpc/grpc-js": "^1.5.7",
 		"common-tags": "^1.8.2",
 		"discord.js": "^13.6.0",
 		"express": "^4.17.2",
@@ -77,7 +79,7 @@
 		"zlib-sync": "^0.1.7"
 	},
 	"volta": {
-		"node": "16.13.2",
+		"node": "16.14.0",
 		"yarn": "1.22.17"
 	},
 	"packageManager": "yarn@3.2.0-rc.11"
